Break Places clients out into separate files, fix one stray `num` left in the 1.2 client.
